Actual ES6 conventions (kinda)
I attempted to change some glaringly obvious things to better ES6 practices.
 - Var to let or const

This shouldn't cause any problems because the scoping looks perfectly fine, but the disadvantage of doing it this way is that it could cause problems depending on context. If this starts causing random undefined things, I recommend changing the problematic declaration to a var. When possible, under the shiny new ES6 conventions, let is the better choice if scope permits it.